+++ lrtable/src/lib/pager.rs
@@ -162,7 +162,7 @@ pub fn pager_stategraph<StorageT: Hash + PrimInt + Unsigned>
     let mut state0 = Itemset::new(grm);
     let mut ctx = Vob::from_elem(grm.terms_len() as usize, false);
     ctx.set(usize::from(grm.eof_term_idx()), true);
-    state0.add(grm.start_prod(), SIdx::from(0 as u32), &ctx);
+    state0.add(grm.start_prod(), SIdx::from(0u32), &ctx);
     closed_states.push(None);
     core_states.push(state0);
     edges.push(HashMap::new());
